envs/syllogym_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "syllogym-env"
version = "0.1.0"
description = "SylloGym: Legal Syllogistic Reasoning Environment for RLVR"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
    "datasets>=2.19.0,<3.0.0",
    "huggingface_hub>=0.23.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0",
]

[project.scripts]
server = "syllogym_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = [
    "syllogym_env",
    "syllogym_env.server",
    "syllogym_env.server.core",
    "syllogym_env.server.generators",
]
[tool.setuptools.package-dir]
"syllogym_env" = "."
"syllogym_env.server" = "server"
"syllogym_env.server.core" = "server/core"
"syllogym_env.server.generators" = "server/generators"
